fix(core): 添加按ID获取用户的接口并重构错误处理
• 新增 UserIDRequest 结构体和通过ID查询用户的接口
• 重构错误变量定义，将 ErrNotFound 移入统一错误集中
• 调整路由注册顺序以支持新的用户查询路径

+++ internal/auth/jwt.go
@@ -32,10 +32,10 @@ type AccessClaims struct {
 }
 
 // GenerateAccessToken 生成 Access Token (访问令牌)
-func GenerateAccessToken(userID uint, role string) (string, error) {
+func GenerateAccessToken(userID uint64, role string) (string, error) {
 	exp := time.Now().UTC().Add(15 * time.Minute) // 24小时有效期
 	claims := AccessClaims{
-		UserID: strconv.FormatUint(uint64(userID), 10),
+		UserID: strconv.FormatUint(userID, 10),
 		Role:   role,
 		RegisteredClaims: jwt.RegisteredClaims{
 			ExpiresAt: jwt.NewNumericDate(exp),
